% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerJsonParser.R
\name{answerJsonParser}
\alias{answerJsonParser}
\title{Removes JSON format from answer options}
\usage{
answerJsonParser(answerOption)
}
\arguments{
\item{answerOption}{A item character answer option object with JSON format.}
}
\value{
Item answer without JSON format.
}
\description{
Removes JSON format from answer options
}
\examples{
answerOption <- paste0("{\\"layout\\":\\"Numpad\\",\\"mediaType\\":\\"",
                   "maximizedTextAnswerOptions\\",\\"answerOptions\\":[\\"0\\"]}")
options <- answerJsonParser(answerOption)
}

