% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemJsonParser.R
\name{itemJsonParser}
\alias{itemJsonParser}
\title{Removes JSON format from item datafile.}
\usage{
itemJsonParser(items)
}
\arguments{
\item{items}{A Oefenweb item dataframe or a Oefenweb item question or Ofenweb
answer options vector.}
}
\value{
Input item dataframe or vector without JSON format.
}
\description{
Removes JSON format from item datafile.
}

