% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionJsonParser.R
\name{questionJsonParser}
\alias{questionJsonParser}
\title{Removes JSON format from question}
\usage{
questionJsonParser(question)
}
\arguments{
\item{question}{A item character question object with JSON format.}
}
\value{
Item question without JSON format.
}
\description{
Removes JSON format from question
}
\examples{
question <- paste0('{"jsonTypeDefinition":"gameSelector","type":"OpenAnswer",',
                   '"question":{"mediaType":"maximizedText","content":"0 x 1"}}')
content <- questionJsonParser(question)$content
type <- questionJsonParser(question)$mediaType
}

