% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oefenwebItemJsonParser.R
\docType{package}
\name{oefenwebItemJsonParser}
\alias{oefenwebItemJsonParser}
\alias{oefenwebItemJsonParser-package}
\title{oefenwebItemJsonParser: A package to remove JSON format}
\description{
The oefenwebItemJsonParser package provides functionality to remove JSON format
from Oefenweb items. The parser can be applied to a full dataset of items to
remove JSON format from the item questions and item answer options. In addition,
JSON can also be removed from only the questions or only the answer options.
}
\section{oefenwebItemJsonParser functions}{
 The oefenwebItemJsonParser package
contains the main function itemJsonParser
which removes JSON format from an item dataset or form a question or
answer options vector. The package also contains a function called questionJsonParser
and answerJsonParser to remove JSON format of a single question or answer options resepctively.
}

